# api.yml
...
openapi: '3.0.0'
info:
  description: TourConnect's Ti2 sync service
  version: 0.0.1 
  title: Ti2
  contact:
    email: engineering@tourconnect.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Message:
      type: object
      properties:
        value:
          type: string
...
    AnyValue:
      description: Can be anything
      type: object
      properties:
        payload:
          type: object
          example: random lorem
    UserApp:
      description: A user + App configuration
      type: object
      properties:
        tokenHint:
          type: string
          description: An identifier for this token, it should NOT be the token itself but a short description of it
          example: 'expediaTestLocation-23'
        token:
          type: object
          description: this is the secret payload that is passed doen to the plugin
          example: { endpoint: 'http://expedia.penguin.local', apiKey: 'this is a secret key' }
    UserAppDelete:
      description: A user + App configuration to delete
      type: object
      properties:
        tokenHint:
          type: string
          description: An identifier for this token, it should NOT be the token itself but a short safe version of it
          example: 'expediaTestLocation-23'
...
    AppTokenList:
      type: object
      properties:
        userAppKeys:
          type: array
          items:
            type: object
            properties:
              integrationId:
                type: string
              userId:
                type: string
              hint:
                type: string
              createdAt:
                type: string
              updatedAt:
                type: string
...
paths:
...
  /app/encode/{app}:
    security:
      - bearerAuth: ['app']
    put:
      tags:
        - app
      summary: Should encode an arbitriary field body into a hash
      security:
        - bearerAuth: ['app']
      operationId: jwtEncode
      parameters:
        - name: app
          in: path
          description: name of the requesting app
          required: true
          example: 'expedia'
          schema:
            type: string
      requestBody:
        description: Payload to encode
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnyValue'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /app/tokens/{app}:
    get:
      security:
        - bearerAuth: ['app']
      tags:
        - app
      summary: List all users/tokens for this app
      operationId: listAppTokens
      parameters:
        - name: app
          in: path
          description: app name to list the users for
          example: 'expedia'
          required: true
          schema:
            type: string
      responses:
        200:
          description: success
          content:
            application/json:
              schema: '#/components/schemas/AppTokenList'
  /{app}/{userId}:
    security:
      - bearerAuth: ['app']
    post:
      tags:
        - app
        - user
        - admin
      summary: Create a user app+Token configuration, the authorization can be either from the app or the user
      security:
        - bearerAuth: ['app', 'user', 'admin']
      operationId: createAppToken
      parameters:
        - name: app
          in: path
          description: app name to create the userId for (must match app token)
          example: 'expedia'
          required: true
          schema:
            type: string
        - name: userId
          in: path
          description: user to create the app for
          example: '536830b6ed19afa44a000002'
          required: true
          schema:
            type: string
      requestBody:
        description: A new User+App configuration
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserApp'
      responses:
        200:
          description: success
          content:
            application/json:
              schema: '#/components/schemas/Message'
    delete:
      tags:
        - app
        - user
      summary: Should delete a user+appToken configuration, the auth token can be from an app or a user
      security:
        - bearerAuth: ['admin', 'app', 'user']
      operationId: deleteAppToken
      parameters:
        - name: app
          in: path
          description: app name to create the userId for (must match app token)
          example: 'expedia'
          required: true
          schema:
            type: string
        - name: userId
          in: path
          description: user to delete the app for
          example: '536830b6ed19afa44a000002'
          required: true
          schema:
            type: string
      requestBody:
        description: The User+App configuration to remove
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAppDelete'
      responses:
        200:
          description: success
          content:
            application/json:
              schema: '#/components/schemas/Message'
...